+++ src/php/Gdbots/Schemas/Analytics/Mixin/TrackedMessage/TrackedMessageV1Trait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Gdbots\Schemas\Analytics\Mixin\TrackedMessage;
+
+use Gdbots\Pbj\Schema;
+
+/**
+ * @method static Schema schema
+ * @method mixed get($fieldName, $default = null)
+ */
+trait TrackedMessageV1Trait
+{
+}